+++ androidApp/src/main/java/com/vorue/pokedex/android/ui/viewModel/PokemonViewModel.kt
@@ -13,7 +13,7 @@ class PokemonViewModel : ViewModel() {
     val pokemon = MutableLiveData<PokedexResults>()
 
 
-    fun setPokemon(id: Int) {
+    fun setPokemon(id: Long) {
        /*viewModelScope.launch(coroutineExceptionHandler) {
             kotlin.runCatching {
                 val pokedexRepository =  PokedexRepository()
